function file created to split some modules
