+++ src/index.js
@@ -1,3 +1,14 @@
+import {
+  getTask,
+  clearElement,
+  renderTaskCount,
+  updateTodo,
+  renderTasks,
+  renderLists,
+  createTask,
+  createList,
+} from './functions';
+
 const listsContainer = document.querySelector('[data-lists]');
 const newListForm = document.querySelector('[data-new-list-form]');
 const newListInput = document.querySelector('[data-new-list-input]');
@@ -6,10 +17,8 @@ const listDisplayContainer = document.querySelector(
   '[data-list-display-container]',
 );
 const listTitleElement = document.querySelector('[data-list-title]');
-const listCountElement = document.querySelector('[data-list-count]');
 const tasksContainer = document.querySelector('[data-tasks]');
-const taskTemplate = document.getElementById('task-template');
-const newTaskForm = document.querySelector('[data-new-task-form]');
+
 const newTaskInput = document.querySelector('[data-new-task-input]');
 const clearCompleteTasksButton = document.querySelector(
   '[data-clear-complete-tasks-button]',
@@ -18,14 +27,40 @@ const priority = document.getElementById('priority');
 const date = document.getElementById('date');
 const description = document.getElementById('description');
 const btnTask = document.getElementById('addTask');
-const allTasks = document.querySelectorAll('.task');
-
-// localstorage creating key
 const LOCAL_STORAGE_LIST_KEY = 'task.lists';
 const LOCAL_STORAGE_SELECTED_LIST_ID_KEY = 'task.selectedListId';
+
 let lists = JSON.parse(localStorage.getItem(LOCAL_STORAGE_LIST_KEY)) || [];
 let selectedListId = localStorage.getItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY);
-// end
+const save = () => {
+  localStorage.setItem(LOCAL_STORAGE_LIST_KEY, JSON.stringify(
+  lists
+  ));
+  localStorage.setItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY,
+    selectedListId);
+};
+
+
+const render = () => {
+  clearElement(listsContainer);
+  renderLists(lists, selectedListId);
+
+  const selectedList = lists.find((list) => list.id === selectedListId);
+  if (selectedListId == null) {
+    listDisplayContainer.style.display = 'none';
+  } else {
+    listDisplayContainer.style.display = '';
+    listTitleElement.innerText = selectedList.name;
+    renderTaskCount(selectedList);
+    clearElement(tasksContainer);
+    renderTasks(selectedList);
+  }
+};
+
+const saveAndRender = () => {
+  save();
+  render();
+};
 
 listsContainer.addEventListener('click', (e) => {
   if (e.target.tagName.toLowerCase() === 'li') {
@@ -65,29 +100,13 @@ tasksContainer.addEventListener('click', (e) => {
   }
 });
 
-function getTask(task) {
-  document.getElementById('newTaskName').value = task.name;
-  document.getElementById('description').value = task.description;
-  document.getElementById('date').value = task.date;
-  document.getElementById('priority').value = task.priority;
-}
-
-function updateTodo(list, task) {
-  task.name = document.getElementById('newTaskName').value;
-  task.description = document.getElementById('description').value;
-  task.date = document.getElementById('date').value;
-  task.priority = document.getElementById('priority').value;
-  const indexTask = list.tasks.indexOf(task);
-  list.tasks[indexTask] = task;
-}
-
-clearCompleteTasksButton.addEventListener('click', (e) => {
+clearCompleteTasksButton.addEventListener('click', () => {
   const selectedList = lists.find((list) => list.id === selectedListId);
   selectedList.tasks = selectedList.tasks.filter((task) => !task.complete);
   saveAndRender();
 });
 
-deleteListButton.addEventListener('click', (e) => {
+deleteListButton.addEventListener('click', () => {
   lists = lists.filter((list) => list.id !== selectedListId);
   selectedListId = null;
   saveAndRender();
@@ -122,93 +141,5 @@ btnTask.addEventListener('click', (e) => {
   selectedList.tasks.push(task);
   saveAndRender();
 });
-
-function createList(name) {
-  return {
-    id: Date.now().toString(),
-    name,
-    tasks: [],
-  };
-}
-
-function createTask(name, description, priority, date) {
-  return {
-    id: Date.now().toString(),
-    name,
-    description,
-    priority,
-    date,
-    complete: false,
-  };
-}
-
-function renderTasks(selectedList) {
-  selectedList.tasks.forEach((task) => {
-    const taskElement = document.importNode(taskTemplate.content, true);
-    const checkbox = taskElement.querySelector('input');
-    checkbox.id = task.id;
-    checkbox.checked = task.complete;
-    const label = taskElement.querySelector('label');
-    label.htmlFor = task.id;
-    label.append(task.name);
-    label.append(task.description);
-    label.append(task.priority);
-    label.append(task.date);
-    tasksContainer.appendChild(taskElement);
-  });
-}
-
-function saveAndRender() {
-  save();
-  render();
-}
-
-function save() {
-  localStorage.setItem(LOCAL_STORAGE_LIST_KEY, JSON.stringify(lists));
-  localStorage.setItem(LOCAL_STORAGE_SELECTED_LIST_ID_KEY, selectedListId);
-}
-
-function render() {
-  clearElement(listsContainer);
-  renderLists();
-
-  const selectedList = lists.find((list) => list.id === selectedListId);
-  if (selectedListId == null) {
-    listDisplayContainer.style.display = 'none';
-  } else {
-    listDisplayContainer.style.display = '';
-    listTitleElement.innerText = selectedList.name;
-    renderTaskCount(selectedList);
-    clearElement(tasksContainer);
-    renderTasks(selectedList);
-  }
-}
-
-
-function renderTaskCount(selectedList) {
-  const incompleteTaskCount = selectedList.tasks.filter((task) => !task.complete)
-    .length;
-  const taskString = incompleteTaskCount === 1 ? 'task' : 'tasks';
-  listCountElement.innerText = `${incompleteTaskCount} ${taskString} remaining`;
-}
-
-function renderLists() {
-  lists.forEach((list) => {
-    const listElement = document.createElement('li');
-    listElement.dataset.listId = list.id;
-    listElement.classList.add('list-name');
-    listElement.innerText = list.name;
-    if (list.id === selectedListId) {
-      listElement.classList.add('active-list');
-    }
-    listsContainer.appendChild(listElement);
-  });
-}
-
-function clearElement(element) {
-  while (element.firstChild) {
-    element.removeChild(element.firstChild);
-  }
-}
 
 render();
